Revert "Merge pull request #175 from fukamachi/swank-port"
This reverts commit 53a93ff396d178dfb41fafb99d23d4446d864351, reversing
changes made to 6fd0279424f7ba5fd4f92d69a1970846b0b11222.

Loading swank as part of this system conflicts with the emacs SLIME
package's own implementation of swank, causing conflicts and a series of
stream errors when developing in emacs. In general, you cannot load two
different versions of swank, and loading the version shipped with the
SLIME package in elpa _and_ the one loaded from quicklisp here causes
that problem.

Swank is an optional nice-to-have feature for people doing certain kinds
of development workflows, but it's not necessary for the library to work
as intended. This could maybe be factored out into a separate system,
but for now remove the swank support so the system can load.